+++ src/test/kotlin/com/github/kindermax/intellijlets/completion/field/FieldsTest.kt
@@ -1,24 +1,29 @@
 package com.github.kindermax.intellijlets.completion.field
 
+import com.github.kindermax.intellijlets.DEFAULT_SHELLS
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import junit.framework.TestCase
 import org.junit.Test
 
 open class FieldsTest : BasePlatformTestCase() {
 
-    override fun getTestDataPath(): String {
-        return "src/test/resources/completion"
-    }
-
     @Test
     fun testRootCompletion() {
-        val letsFile = myFixture.copyFileToProject("/root/lets.yaml")
-        myFixture.configureFromExistingVirtualFile(letsFile)
-        val variants = myFixture.getCompletionVariants("/root/lets.yaml")
+        myFixture.configureByText(
+            "lets.yaml",
+            """
+            shell: bash
+    
+            commands:
+              run:
+                cmd: Echo Run
+            
+            <caret>
+            """.trimIndent()
+        )
+        val variants = myFixture.getCompletionVariants("lets.yaml")
             ?: return TestCase.fail("completion variants must not be null")
         val expected = listOf(
-            "shell",
-            "commands",
             "env",
             "eval_env",
             "version",
@@ -30,10 +35,31 @@ open class FieldsTest : BasePlatformTestCase() {
     }
 
     @Test
+    fun testShellCompletion() {
+        myFixture.configureByText(
+            "lets.yaml",
+            """
+            shell: <caret>
+            """.trimIndent()
+        )
+        val variants = myFixture.getCompletionVariants("lets.yaml")
+            ?: return TestCase.fail("completion variants must not be null")
+        TestCase.assertEquals(DEFAULT_SHELLS.sorted(), variants.sorted())
+    }
+
+    @Test
     fun testCommandCompletionWithCLetter() {
-        val letsFile = myFixture.copyFileToProject("/command/lets.yaml")
-        myFixture.configureFromExistingVirtualFile(letsFile)
-        val variants = myFixture.getCompletionVariants("/command/lets.yaml")
+        myFixture.configureByText(
+            "lets.yaml",
+            """
+            shell: bash
+            commands:
+              echo:
+                c<caret>
+            """.trimIndent()
+        )
+
+        val variants = myFixture.getCompletionVariants("lets.yaml")
             ?: return TestCase.fail("completion variants must not be null")
         val expected = listOf(
             "description",
@@ -46,10 +72,81 @@ open class FieldsTest : BasePlatformTestCase() {
     }
 
     @Test
+    fun testCommandOptionsCompletionOnlyOptions() {
+        myFixture.configureByText(
+            "lets.yaml",
+            """
+            shell: bash
+            commands:
+              echo:
+                op<caret>
+            """.trimIndent()
+        )
+
+        myFixture.completeBasic()
+
+        TestCase.assertEquals(myFixture.caretOffset, 63)
+        TestCase.assertEquals(
+            myFixture.file.text.trimIndent(),
+            """
+            shell: bash
+            commands:
+              echo:
+                options: |
+                  Usage: lets 
+            """.trimIndent()
+        )
+    }
+
+    @Test
+    fun testCommandOptionsCompletionNotOnlyOptions() {
+        myFixture.configureByText(
+            "lets.yaml",
+            """
+            shell: bash
+            commands:
+              echo:
+                cmd: echo Hi
+                op<caret>
+            """.trimIndent()
+        )
+
+        myFixture.completeBasic()
+
+        TestCase.assertEquals(myFixture.caretOffset, 80)
+        TestCase.assertEquals(
+            myFixture.file.text.trimIndent(),
+            """
+            shell: bash
+            commands:
+              echo:
+                cmd: echo Hi
+                options: |
+                  Usage: lets 
+            """.trimIndent()
+        )
+    }
+
+    @Test
     fun testDependsCompletionWorks() {
-        val letsFile = myFixture.copyFileToProject("/depends/lets.yaml")
-        myFixture.configureFromExistingVirtualFile(letsFile)
-        val variants = myFixture.getCompletionVariants("/depends/lets.yaml")
+        myFixture.configureByText(
+            "lets.yaml",
+            """
+            shell: bash
+            commands:
+              hello:
+                cmd: echo Hello
+                depends: [world]
+              hi:
+                cmd: echo Hi
+              lol:
+                cmd: echo Lol
+              world:
+                depends: [<caret>]
+                cmd: echo World
+            """.trimIndent()
+        )
+        val variants = myFixture.getCompletionVariants("lets.yaml")
             ?: return TestCase.fail("completion variants must not be null")
 
         val expected = listOf(
@@ -57,5 +154,53 @@ open class FieldsTest : BasePlatformTestCase() {
         )
 
         TestCase.assertEquals(expected.sorted(), variants.sorted())
+    }
+
+    @Test
+    fun testDependsCompletionWorksNoCommandYet() {
+        myFixture.configureByText(
+            "lets.yaml",
+            """
+            shell: bash
+            commands:
+              hi:
+                cmd: echo Hi
+              run:
+                depends:
+                  - <caret>
+            """.trimIndent()
+        )
+        val variants = myFixture.getCompletionVariants("lets.yaml")
+            ?: return TestCase.fail("completion variants must not be null")
+
+        val expected = listOf("hi")
+
+        TestCase.assertEquals(expected.sorted(), variants.sorted())
+    }
+
+    @Test
+    fun testCommandDependsCompletionArray() {
+        myFixture.configureByText(
+            "lets.yaml",
+            """
+            shell: bash
+            commands:
+              echo:
+                dep<caret>
+            """.trimIndent()
+        )
+
+        myFixture.completeBasic()
+
+        TestCase.assertEquals(
+            myFixture.file.text,
+            """
+            shell: bash
+            commands:
+              echo:
+                depends:
+                  - 
+            """.trimIndent()
+        )
     }
 }