+++ CHANGELOG.md
@@ -3,9 +3,14 @@
 # intellij-lets Changelog
 
 ## [Unreleased]
-## [0.0.6]
+### Added
+- add some default completions for `shell`
+- enhance completions for yaml keys (add :, new line, multiline)
+- add more tests
 ### Updated
 - drop support for 193 IDE version
+- refactor completion provider
+- refactor config, parse gracefully
 
 ## [0.0.6]
 ### Updated